rename rest_api module to http_server
