+++ wazo_dird/plugins/api/http.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -9,7 +9,7 @@ from flask import make_response
 from xivo.chain_map import ChainMap
 from xivo.http_helpers import reverse_proxy_fix_api_spec
 
-from wazo_dird.rest_api import ErrorCatchingResource
+from wazo_dird.http_server import ErrorCatchingResource
 
 logger = logging.getLogger(__name__)
