+++ wazo_dird/plugins/default_json/http.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -19,7 +19,7 @@ from wazo_dird.exception import (
 )
 from wazo_dird.helpers import DisplayAwareResource
 from wazo_dird.auth import required_acl
-from wazo_dird.rest_api import LegacyAuthResource, AuthResource
+from wazo_dird.http_server import LegacyAuthResource, AuthResource
 
 logger = logging.getLogger(__name__)
