+++ wazo_dird/plugins/google_backend/http.py
@@ -1,4 +1,4 @@
-# Copyright 2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2019-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -6,7 +6,7 @@ import logging
 from flask import request
 from wazo_dird.auth import required_acl
 from wazo_dird.helpers import SourceItem, SourceList
-from wazo_dird.rest_api import AuthResource
+from wazo_dird.http_server import AuthResource
 from xivo.tenant_flask_helpers import Tenant, token
 
 from .schemas import contact_list_schema, list_schema, source_list_schema, source_schema
