+++ wazo_dird/plugins/phonebook/http.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -22,7 +22,7 @@ from wazo_dird.exception import (
     NoSuchPhonebook,
     NoSuchTenant,
 )
-from wazo_dird.rest_api import LegacyAuthResource
+from wazo_dird.http_server import LegacyAuthResource
 
 logger = logging.getLogger(__name__)
