+++ wazo_dird/plugins/wazo_user_backend/http.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2018-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import request
@@ -7,7 +7,7 @@ from xivo.tenant_flask_helpers import Tenant
 
 from wazo_dird.auth import required_acl
 from wazo_dird.helpers import SourceItem, SourceList
-from wazo_dird.rest_api import AuthResource
+from wazo_dird.http_server import AuthResource
 from wazo_dird.plugin_helpers.confd_client_registry import registry
 
 from .contact import ContactLister
